libvirt: add support for qemu_group_auto to infer your qemu group
libvirt will be used by Linux distributions, these use qemu. The qemu
group is currently set by either a variable on the configuration or
set by an environment variable. We want to be able to easily allow
deployment of virtualization solutions without much work on the target
host. Add support to enable inference of your qemu group based on your
Linux distribution.

If you want to enable this by default on your project you can set
the global libvirt config qemu_group_auto variable to true. Otherwise
set this to false.

The environment variable for the project *always* overrides this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>