lacks_nvme_support: add server specific setting
If we are booting a system with no nvme drive support, we want
to use an alternative. In this case we can just tell qemu to use
whatever drive it prefers, but we still prefer virtio.

Later on we can add another knob to disable virtio for systems
which lack support for it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>