Vagrantfile: allow to skip_ansible
Some old hosts on with python older than 2.6 won't be able
to work when your server is using a recent version of ansible.
What you need to do is at least install python 2.6, but this
requires some work on your hosts.

Allow the user to specify which hosts to bypass ansible on.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>