Vagrantfile: support configuration overrides in *_nodes_override.yaml
Often times you may want to just bump the amount of memory, or
drives, but not have to deal with this data in your project
being part of the visible data on git diff. To let you override
data you can use a file with postfix _nodes_override.yaml for
your configuration.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>