Vagrantfile: add support to specify python interpreter
Using ansible to update the cache shows an error saying
that python-apt is not available or installed. This is
because the old interpreter is beind auto detected.

To force the new one we must specify this in ansible.
There do this through the hosts file in ansible, but
when Vagrant uses ansible we need to specify this
thorugh an extra variable.

Since we expect some hosts will require the old interpreter
we then enable support to specify an interpreter globally, and
also per host.

The vagrant ansible portion covers only ansible playbooks
ran through provisioning, if you enabled that. If you don't
use ansible provisioning for some playbooks you don't
have to do anything.

Then, on top of this, if you *do* run ansible manually
you'll want to update your own hosts file with something
like this:

[all:vars]
ansible_python_interpreter =  "/usr/bin/python3"

If unsure, just look at the kdevops project which
will be updated to take advantage of this new role update
and also update its own hosts file respectively.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>