Vagrantfile: fix and document support for extra_args
How we use the extra_args for ansible has so far been undocumented.
We want to properly support this feature and to ensure it doesn't clash
with our other features.

We used to support just using the extra_args stuff by letting you
specify in your node configuration something like this:

ansible_playbooks:
  # If this file exists you can override any ansible variable there.
  # This file is optional.
  extra_vars: "../extra_vars.yml"
  playbooks:
    - name: "../playbooks/update_ssh_config_vagrant.yml"
    - name: "../playbooks/devconfig.yml"

We have *always* set this extra_vars variable up, and so if it was
found, Vagrantfile would ensure vagrant processes it.

The way the vagrant ansible plugin works though is to require you
to set the ansible.extra_vars and this can be set as either the
file, or a hash. Unfortunately the documentation on vagrant about
what type of hash to use is not clear. I checked the upstream code
and the hash needs to be in JSON format. Also, the ansible plugin
doesn't do any of its own processing of the file if a file is passed,
it leaves this to ansible to figure out what to do with the file.

If a file is passed, all that the vagrant ansible plugin is doing is
expanding the command line with:

  --extra-vars=file

Now, we recently added support for letting you specify the python
interpreter, and this is required given otherwise ansible won't work
on some newer systems. Ansible uses an old interpreter by default
to ensure backward compatibility with older systems.

The way we inform ansible about our pythion interpreter preference
though is by using a JSON hash and stuffing it into the ansible.extra_vars.
Now, since we typically never used the ansible_playbooks.extra_vars in
projects, we currently know we have tested when this file is declared
but it doesn't exist. Since we are setting the interpreter in the
kdevops demo project it means we have also tested the python interpreter
code path as well.

What was broken was when a user actually had the ../extra_vars.yml file
present as a real file, *and* the python interpreter was passed. In such
case we would pass on the python interpterer as a JSON hash followed by
the file name...

Fix this by actually processing the file and giving the hash value of
the file. This means we must check the file extension also. Since
vagrant only processes hash values as JSON, we must then convert
yaml to a JSON hash and append that, and if a json file is used, process
the json file to a hash value and append that.

If no python interpreter is passed we just give the vagrant ansible
plugin the file as we used to, so that it can then call ansible with the

  --extra-vars=file

So processing of the file is left to ansible. Lastly, when the desire is
to pass a file to ansible with extra args, we must prefix the string
with the character "@", and so processing of the file was always broken.

Fixes: bd80ce49cf31 ("Vagrantfile: add support to specify python interpreter")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>