Vagrantfile: simplify python interpreter settings by reading inventory file
It turns out that if we define a python interpreter per hosts, only the
last one takes effect for all systems. This is partly due to the way we
handle playbooks and ensure to only run once that loop.

We can simplify things considerably by just reading the inventory file,
given that you are supposed to have this file anyway. If present we
rely on it, and by default assume you have it on ../hosts. You can set
this on the configuration for ansible as well.

This also lets us keep the way we use the extra_vars simple, by just
passing the argument to ansible, instead of doing parsing ourselves to
combine with the python interpreter JSON hash we created.

Fixes: bd80ce49cf31 ("Vagrantfile: add support to specify python interpreter")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>