tasks: force deploying Vagrantfile
It more common to run into the issue that you have upgraded
kdevops, and you wonder why something is not working, only to
realize kdevops was being too careful and not upgrading your
Vagrantfile for you if it already existed.

This clearly just silly now.

Yes, we risk overriding a file you were working on, but tough luck.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>